requirements.txt
django==3.2.9
mysqlclient==2.1.0
django-debug-toolbar~=3.2.4
pillow
djangorestframework-jwt
django-cors-headers